Reworked Income
Reworked income, and am not 100% sure this is the right way. The idea
was to accommodate buildings like camp, which produce more than one
resource. Also hopefully makes income a little more stable/robust and
easy to modify by making it based on number of structures instead of
whatever income was prior. Hoping this approach will save time in the
long run.